setup snapshot testing
