+++ protocols/src/ir/meta_graphviz.rs
@@ -1,0 +1,97 @@
+use crate::ir::meta_automaton::{MetaAutomaton, MetaFrontier, MetaOutcome, MetaState};
+
+fn format_state(graph: &MetaAutomaton, state: &MetaState) -> String {
+    let mut parts: Vec<String> = state
+        .live
+        .iter()
+        .map(|transaction| {
+            format!(
+                "{}<SUB>{}</SUB><SUP>({})</SUP>",
+                escape_html(&graph.protocols[transaction.protocol].name),
+                transaction.post_cycle,
+                transaction.instance
+            )
+        })
+        .collect();
+
+    if let Some(frontier) = state.frontier {
+        parts.push(match frontier {
+            MetaFrontier::Choice { instance } => format!("E<SUP>({instance})</SUP>"),
+            MetaFrontier::Pre {
+                protocol,
+                instance,
+                elapsed,
+            } => {
+                let phase = if graph.protocols[protocol].pre_cycles.is_none() {
+                    "pre".to_string()
+                } else {
+                    format!("pre[{elapsed}]")
+                };
+                format!(
+                    "{}<SUB>{}</SUB><SUP>({})</SUP>",
+                    escape_html(&graph.protocols[protocol].name),
+                    phase,
+                    instance
+                )
+            }
+        });
+    }
+
+    if parts.is_empty() {
+        "DONE".to_string()
+    } else {
+        parts.join(" ∧ ")
+    }
+}
+
+fn escape_html(text: &str) -> String {
+    text.replace('&', "&amp;")
+        .replace('<', "&lt;")
+        .replace('>', "&gt;")
+}
+
+pub fn to_dot(graph: &MetaAutomaton) -> String {
+    let mut output = String::from(
+        "digraph \"driver_meta\" {\n  rankdir=LR;\n  node [shape=box];\n  entry [shape=plain,label=\"ENTRY\"];\n",
+    );
+    output.push_str(&format!("  entry -> node{};\n", graph.entry));
+
+    for (node_id, node) in graph.nodes.iter().enumerate() {
+        output.push_str(&format!(
+            "  node{node_id} [label=<{}>];\n",
+            format_state(graph, &node.state)
+        ));
+        for edge in &node.edges {
+            let outcome = match edge.outcome {
+                MetaOutcome::Fork => "forks",
+                MetaOutcome::Continue => "continues",
+            };
+            let mut label = format!(
+                "{} {outcome}",
+                escape_html(&graph.protocols[edge.protocol].name)
+            );
+            if !edge.rotations.is_empty() {
+                let rotations = edge
+                    .rotations
+                    .iter()
+                    .map(|rotation| {
+                        let name = escape_html(&graph.protocols[rotation.protocol].name);
+                        if rotation.amount == 1 {
+                            format!("R<SUB>{name}</SUB>")
+                        } else {
+                            format!("R<SUB>{name}</SUB><SUP>{}</SUP>", rotation.amount)
+                        }
+                    })
+                    .collect::<Vec<_>>()
+                    .join(", ");
+                label.push_str(&format!(" / {rotations}"));
+            }
+            output.push_str(&format!(
+                "  node{node_id} -> node{} [label=<{}>];\n",
+                edge.target, label
+            ));
+        }
+    }
+    output.push_str("}\n");
+    output
+}
